requirements.txt
asgiref==3.6.0
Django==4.1.1
sqlparse==0.4.3
python-dotenv==0.21.1
psycopg2==2.9.5
stripe==5.1.1
